+++ SG.TestRunService/Migrations/TSDbContextModelSnapshot.cs
@@ -106,8 +106,8 @@ namespace SG.TestRunService.Migrations
                         .HasColumnType("nvarchar(50)")
                         .HasMaxLength(50);
 
-                    b.Property<int>("Type")
-                        .HasColumnType("int");
+                    b.Property<byte>("Type")
+                        .HasColumnType("tinyint");
 
                     b.HasKey("Id");
